[Feat] 공통 StationDetailBody·카카오식 IA·검색/즐겨찾기 연결
Refs #2436
